+++ BUILD.gn
@@ -542,6 +542,9 @@ config("grpc_config") {
         "src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.h",
         "src/core/ext/upb-generated/validate/validate.upb.c",
         "src/core/ext/upb-generated/validate/validate.upb.h",
+        "src/core/ext/xds/certificate_provider_factory.h",
+        "src/core/ext/xds/certificate_provider_registry.h",
+        "src/core/ext/xds/certificate_provider_store.h",
         "src/core/ext/xds/xds_api.cc",
         "src/core/ext/xds/xds_api.h",
         "src/core/ext/xds/xds_bootstrap.cc",
@@ -794,11 +797,7 @@ config("grpc_config") {
         "src/core/lib/security/authorization/evaluate_args.h",
         "src/core/lib/security/authorization/mock_cel/activation.h",
         "src/core/lib/security/authorization/mock_cel/cel_value.h",
-        "src/core/lib/security/certificate_provider/config.h",
-        "src/core/lib/security/certificate_provider/factory.h",
-        "src/core/lib/security/certificate_provider/provider.h",
-        "src/core/lib/security/certificate_provider/registry.h",
-        "src/core/lib/security/certificate_provider/store.h",
+        "src/core/lib/security/certificate_provider.h",
         "src/core/lib/security/context/security_context.cc",
         "src/core/lib/security/context/security_context.h",
         "src/core/lib/security/credentials/alts/alts_credentials.cc",
