+++ gRPC-C++.podspec
@@ -369,6 +369,9 @@ Pod::Spec.new do |s|
                       'src/core/ext/upb-generated/udpa/annotations/versioning.upb.h',
                       'src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.h',
                       'src/core/ext/upb-generated/validate/validate.upb.h',
+                      'src/core/ext/xds/certificate_provider_factory.h',
+                      'src/core/ext/xds/certificate_provider_registry.h',
+                      'src/core/ext/xds/certificate_provider_store.h',
                       'src/core/ext/xds/xds_api.h',
                       'src/core/ext/xds/xds_bootstrap.h',
                       'src/core/ext/xds/xds_channel.h',
@@ -518,11 +521,7 @@ Pod::Spec.new do |s|
                       'src/core/lib/security/authorization/evaluate_args.h',
                       'src/core/lib/security/authorization/mock_cel/activation.h',
                       'src/core/lib/security/authorization/mock_cel/cel_value.h',
-                      'src/core/lib/security/certificate_provider/config.h',
-                      'src/core/lib/security/certificate_provider/factory.h',
-                      'src/core/lib/security/certificate_provider/provider.h',
-                      'src/core/lib/security/certificate_provider/registry.h',
-                      'src/core/lib/security/certificate_provider/store.h',
+                      'src/core/lib/security/certificate_provider.h',
                       'src/core/lib/security/context/security_context.h',
                       'src/core/lib/security/credentials/alts/alts_credentials.h',
                       'src/core/lib/security/credentials/alts/check_gcp_environment.h',
@@ -868,6 +867,9 @@ Pod::Spec.new do |s|
                               'src/core/ext/upb-generated/udpa/annotations/versioning.upb.h',
                               'src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.h',
                               'src/core/ext/upb-generated/validate/validate.upb.h',
+                              'src/core/ext/xds/certificate_provider_factory.h',
+                              'src/core/ext/xds/certificate_provider_registry.h',
+                              'src/core/ext/xds/certificate_provider_store.h',
                               'src/core/ext/xds/xds_api.h',
                               'src/core/ext/xds/xds_bootstrap.h',
                               'src/core/ext/xds/xds_channel.h',
@@ -1017,11 +1019,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/security/authorization/evaluate_args.h',
                               'src/core/lib/security/authorization/mock_cel/activation.h',
                               'src/core/lib/security/authorization/mock_cel/cel_value.h',
-                              'src/core/lib/security/certificate_provider/config.h',
-                              'src/core/lib/security/certificate_provider/factory.h',
-                              'src/core/lib/security/certificate_provider/provider.h',
-                              'src/core/lib/security/certificate_provider/registry.h',
-                              'src/core/lib/security/certificate_provider/store.h',
+                              'src/core/lib/security/certificate_provider.h',
                               'src/core/lib/security/context/security_context.h',
                               'src/core/lib/security/credentials/alts/alts_credentials.h',
                               'src/core/lib/security/credentials/alts/check_gcp_environment.h',
