+++ grpc.gemspec
@@ -446,6 +446,9 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/upb-generated/udpa/data/orca/v1/orca_load_report.upb.h )
   s.files += %w( src/core/ext/upb-generated/validate/validate.upb.c )
   s.files += %w( src/core/ext/upb-generated/validate/validate.upb.h )
+  s.files += %w( src/core/ext/xds/certificate_provider_factory.h )
+  s.files += %w( src/core/ext/xds/certificate_provider_registry.h )
+  s.files += %w( src/core/ext/xds/certificate_provider_store.h )
   s.files += %w( src/core/ext/xds/xds_api.cc )
   s.files += %w( src/core/ext/xds/xds_api.h )
   s.files += %w( src/core/ext/xds/xds_bootstrap.cc )
@@ -767,11 +770,7 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/security/authorization/evaluate_args.h )
   s.files += %w( src/core/lib/security/authorization/mock_cel/activation.h )
   s.files += %w( src/core/lib/security/authorization/mock_cel/cel_value.h )
-  s.files += %w( src/core/lib/security/certificate_provider/config.h )
-  s.files += %w( src/core/lib/security/certificate_provider/factory.h )
-  s.files += %w( src/core/lib/security/certificate_provider/provider.h )
-  s.files += %w( src/core/lib/security/certificate_provider/registry.h )
-  s.files += %w( src/core/lib/security/certificate_provider/store.h )
+  s.files += %w( src/core/lib/security/certificate_provider.h )
   s.files += %w( src/core/lib/security/context/security_context.cc )
   s.files += %w( src/core/lib/security/context/security_context.h )
   s.files += %w( src/core/lib/security/credentials/alts/alts_credentials.cc )
