--- src/core/lib/security/certificate_provider/registry.h
@@ -1,56 +1,0 @@
-//
-//
-// Copyright 2020 gRPC authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-//
-
-#ifndef GRPC_CORE_LIB_SECURITY_CERTIFICATE_PROVIDER_REGISTRY_H
-#define GRPC_CORE_LIB_SECURITY_CERTIFICATE_PROVIDER_REGISTRY_H
-
-#include <grpc/support/port_platform.h>
-
-#include <memory>
-#include <unordered_map>
-
-namespace grpc_core {
-
-class CertificateProviderFactory;
-
-// Global registry for all the certificate provider plugins.
-class CertificateProviderRegistry {
- public:
-  // Global initialization of the registry.
-  static void InitRegistry();
-  // Global shutdown of the registry.
-  static void ShutdownRegistry();
-  // Register a provider with the registry. Can only be called after calling
-  // InitRegistry(). The key of the factory is extracted from factory parameter
-  // with method CertificateProviderFactory::name. If the same key is registered
-  // twice, an exception is raised.
-  static void RegisterProvider(
-      std::unique_ptr<CertificateProviderFactory> factory);
-
-  // Returns the factory for the plugin keyed by name.
-  static CertificateProviderFactory* GetFactory(const std::string& name);
-
- private:
-  std::unordered_map<std::string /* plugin name */,
-                     std::unique_ptr<CertificateProviderFactory>>
-      registry_;
-};
-
-}  // namespace grpc_core
-
-#endif  // GRPC_CORE_LIB_SECURITY_CERTIFICATE_PROVIDER_REGISTRY_H
